Format code of task_generator_gui
